# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CloseTerminalRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  handler.NewSessionRequest:
    properties:
      name:
        type: string
    type: object
  handler.NewTerminalRequest:
    properties:
      cols:
        type: integer
      cwd:
        type: string
      name:
        type: string
      rows:
        type: integer
    type: object
...
  handler.RemoveSessionRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  handler.SaveSessionRequest:
    properties:
      id:
        type: string
      messages:
        type: string
      name:
        type: string
    required:
    - id
    type: object
  handler.Session:
    properties:
      created_at:
        type: integer
      id:
        type: string
      messages:
        type: string
      name:
        type: string
      updated_at:
        type: integer
    type: object
  handler.SessionInfo:
    properties:
      created_at:
        type: integer
      id:
        type: string
      name:
        type: string
      updated_at:
        type: integer
    type: object
  handler.SetRequest:
    properties:
      key:
        type: string
      value:
        type: string
    required:
    - key
    - value
    type: object
  handler.TerminalInfo:
    properties:
      cols:
        type: integer
      created_at:
        type: integer
      cwd:
        type: string
      exit_code:
        type: integer
      history_size:
        type: integer
      id:
        type: string
      name:
        type: string
      pty_status:
        type: string
      rows:
        type: integer
      shell:
        type: string
      status:
        type: string
      updated_at:
        type: integer
    type: object
...
host: localhost:1984
info:
  contact: {}
  description: VibeGo 后端服务 API
  title: VibeGo API
  version: 0.0.1
paths:
...
  /api/session/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handler.SessionInfo'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all sessions
      tags:
      - Session
  /api/session/load:
    get:
      parameters:
      - description: Session ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Session'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Load session by ID
      tags:
      - Session
  /api/session/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: New session request
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.NewSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new session
      tags:
      - Session
  /api/session/rm:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Remove session request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RemoveSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove session
      tags:
      - Session
  /api/session/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: Save session request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SaveSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Save session
      tags:
      - Session
  /api/settings/get:
    get:
      parameters:
      - description: Setting key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a setting by key
      tags:
      - Settings
  /api/settings/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all settings
      tags:
      - Settings
  /api/settings/reset:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset all settings
      tags:
      - Settings
  /api/settings/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Set request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set a setting
      tags:
      - Settings
  /api/terminal/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: Terminal ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CloseTerminalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Close terminal session
      tags:
      - Terminal
  /api/terminal/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handler.TerminalInfo'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List terminal sessions
      tags:
      - Terminal
  /api/terminal/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: Terminal options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.NewTerminalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new terminal session
      tags:
      - Terminal
  /api/terminal/ws/{id}:
    get:
      parameters:
      - description: Terminal ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Connect to terminal websocket
      tags:
      - Terminal
...
swagger: "2.0"